various
git-svn-id: https://svn.code.sf.net/p/triplea/code/trunk/triplea@462 a4c711aa-641f-0410-879d-9b464e52b682

+++ src/games/strategy/net/GUID.java
@@ -84,7 +84,7 @@ public class GUID implements Externalizable
 	
 	public int hashCode()
 	{
-		return ((int) m_id) ^  m_prefix.hashCode();
+		return  m_id ^  m_prefix.hashCode();
 	}
 	
 	public String toString()
